fixed casting bug

+++ arduino/firmware/lib/kinematics/Kinematics.cpp
@@ -132,5 +132,5 @@ Kinematics::velocities Kinematics::getVelocities(int motor1, int motor2, int mot
 int Kinematics::rpmToPWM(int rpm)
 {
   //remap scale of target RPM vs MAX_RPM to PWM
-  return (double)((rpm / max_rpm_) * 255);
+ return (((double) rpm / (double) max_rpm_) * 255);
 }
